+++ packages/terra-graphs/tests/wdio/Graphs/ScatterGraph/scatter-graph-spec.js
@@ -1,5 +1,5 @@
 Terra.describeViewports('ScatterGraph ', ['tiny', 'medium', 'large'], () => {
-  describe('Basic Satter graphs', () => {
+  describe('Basic Scatter graphs', () => {
     it('validates simple scatter graph container', () => {
       browser.url('/raw/tests/terra-graphs-docs/graphs/scatter-graph/basic-scatter-graphs/simple-scatter');
       Terra.validates.screenshot('simple_scatter', { selector: '.carbon-graph-container' });