+++ packages/terra-graphs/CHANGELOG.md
@@ -11,5 +11,6 @@
   * Created component and wdio test cases for the Timeline graph.
   * Created component and wdio test cases for the Pie chart.
   * Created component and wdio test cases for the Combination graph.
+  * Added CSS and new example tests for timeline graph.
 
 * Initial stable release